+++ pyproject.toml
@@ -14,6 +14,25 @@ pygithub = "^1.58.2"
 configparser = "^5.3.0"
 
 
+[tool.poetry.group.dev.dependencies]
+flake8 = "^6.0.0"
+black = "^23.3.0"
+mypy = "^1.3.0"
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.black]
+line-length = 120
+target_version = ["py310"]
+
+[tool.isort]
+profile = "black"
+line_length = 120
+
+[tool.flake8]
+ignore = ["E501"]
+
+[tool.mypy]
+ignore_missing_imports = true
